refactor JavaScript files structure - moved JavaScript files from src/main/assets/js to src/main/js and updated import paths
