+++ src/main/js/logger.ts
