+++ src/main/js/upload-documents.ts
